@@ -34,11 +34,13 @@
     "@typescript-eslint/eslint-plugin": "4.15.1",
     "@typescript-eslint/parser": "4.15.1",
     "earljs": "^0.1.10",
-    "eslint": "^6.8.0",
-    "eslint-config-typestrict": "^1.0.0",
+    "eslint": "^7.29.0",
+    "eslint-config-typestrict": "^1.0.1",
     "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-no-only-tests": "^2.4.0",
-    "eslint-plugin-sonarjs": "^0.5.0",
+    "eslint-plugin-no-only-tests": "^2.6.0",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "eslint-plugin-sonarjs": "^0.8.0-125",
+    "eslint-plugin-unused-imports": "^1.1.1",
     "mocha": "^8.2.0",
     "prettier": "^2.1.2",
     "solc": "^0.6.4",
