@@ -1,10 +1,10 @@
-import { typedAssert, q18, IsExact, AssertTrue } from 'test-utils'
-import { BigNumber, formatBytes32String } from 'ethers/utils'
-import { Awaited } from 'ts-essentials'
 import { expect } from 'earljs'
+import { BigNumber, formatBytes32String } from 'ethers/utils'
+import { AssertTrue, IsExact, q18, typedAssert } from 'test-utils'
+import { Awaited } from 'ts-essentials'
 
-import { createNewBlockchain, deployContract } from './common'
 import { DataTypesInput } from '../types/DataTypesInput'
+import { createNewBlockchain, deployContract } from './common'
 
 describe('DataTypesInput', () => {
   let contract!: DataTypesInput
@@ -77,11 +77,13 @@ describe('DataTypesInput', () => {
   // NOTE: typesAssert is too simple to tests type compatibility here so we can't use it
   it('generates correct types for tuples', () => {
     type ViewTupleType = Awaited<ReturnType<typeof contract.input_tuple>>
+    //eslint-disable-next-line @typescript-eslint/no-unused-vars
     type _t1 = AssertTrue<IsExact<ViewTupleType, [BigNumber, BigNumber]>>
   })
 
   it('generates correct types for structs', () => {
     type ViewStructType = Awaited<ReturnType<typeof contract.input_struct>>
+    //eslint-disable-next-line @typescript-eslint/no-unused-vars
     type _t1 = AssertTrue<
       IsExact<ViewStructType, [BigNumber, BigNumber] & { uint256_0: BigNumber; uint256_1: BigNumber }>
     >
