@@ -1,8 +1,8 @@
+import { BigNumber } from 'ethers/utils'
 import { typedAssert } from 'test-utils'
 
-import { createNewBlockchain, deployContract } from './common'
 import { Overloads } from '../types/Overloads'
-import { BigNumber } from 'ethers/utils'
+import { createNewBlockchain, deployContract } from './common'
 
 describe('Overloads', () => {
   let contract: Overloads
