@@ -7,10 +7,11 @@ import {
   EventDeclaration,
   FunctionDeclaration,
 } from 'typechain'
-import { generateInputType, generateInputTypes } from './types'
-import { codegenFunctions } from './functions'
+
 import { FACTORY_POSTFIX } from '../common'
+import { codegenFunctions } from './functions'
 import { reservedKeywords } from './reserved-keywords'
+import { generateInputType, generateInputTypes } from './types'
 
 export function codegenContractTypings(contract: Contract) {
   const template = `
@@ -76,10 +77,11 @@ export function codegenContractTypings(contract: Contract) {
 
 export function codegenContractFactory(contract: Contract, abi: any, bytecode?: BytecodeWithLinkReferences): string {
   const constructorArgs =
-    (contract.constructor && contract.constructor[0] ? generateInputTypes(contract.constructor[0].inputs) : '') +
+    (contract.constructor[0] ? generateInputTypes(contract.constructor[0].inputs) : '') +
     'overrides?: TransactionOverrides'
-  const constructorArgNamesWithoutOverrides =
-    contract.constructor && contract.constructor[0] ? generateParamNames(contract.constructor[0].inputs) : ''
+  const constructorArgNamesWithoutOverrides = contract.constructor[0]
+    ? generateParamNames(contract.constructor[0].inputs)
+    : ''
   const constructorArgNames = constructorArgNamesWithoutOverrides
     ? `${constructorArgNamesWithoutOverrides}, overrides`
     : 'overrides'
