@@ -8,6 +8,7 @@ import {
   EventDeclaration,
   FunctionDeclaration,
 } from 'typechain'
+
 import { FACTORY_POSTFIX } from '../common'
 import { codegenFunctions } from './functions'
 import { reservedKeywords } from './reserved-keywords'
@@ -140,14 +141,15 @@ export function codegenContractTypings(contract: Contract, codegenConfig: Codege
 
 export function codegenContractFactory(contract: Contract, abi: any, bytecode?: BytecodeWithLinkReferences): string {
   const constructorArgs =
-    (contract.constructor && contract.constructor[0] ? generateInputTypes(contract.constructor[0].inputs) : '') +
+    (contract.constructor[0] ? generateInputTypes(contract.constructor[0].inputs) : '') +
     `overrides?: ${
       contract.constructor[0]?.stateMutability === 'payable'
         ? 'PayableOverrides & { from?: string | Promise<string> }'
         : 'Overrides & { from?: string | Promise<string> }'
     }`
-  const constructorArgNamesWithoutOverrides =
-    contract.constructor && contract.constructor[0] ? generateParamNames(contract.constructor[0].inputs) : ''
+  const constructorArgNamesWithoutOverrides = contract.constructor[0]
+    ? generateParamNames(contract.constructor[0].inputs)
+    : ''
   const constructorArgNames = constructorArgNamesWithoutOverrides
     ? `${constructorArgNamesWithoutOverrides}, overrides || {}`
     : 'overrides || {}'
