@@ -1,8 +1,9 @@
-import { join, resolve, basename } from 'path'
-import { uniqBy, compact } from 'lodash'
+import { compact, uniqBy } from 'lodash'
+import { basename, join, resolve } from 'path'
 import { Dictionary } from 'ts-essentials'
 import {
   BytecodeWithLinkReferences,
+  CodegenConfig,
   Config,
   Contract,
   extractAbi,
@@ -11,15 +12,14 @@ import {
   FileDescription,
   getFileExtension,
   getFilename,
+  normalizeName,
   parse,
   TypeChainTarget,
-  normalizeName,
-  CodegenConfig,
 } from 'typechain'
 
 import { codegenAbstractContractFactory, codegenContractFactory, codegenContractTypings } from './codegen'
-import { FACTORY_POSTFIX } from './common'
 import { generateHardhatHelper } from './codegen/hardhat'
+import { FACTORY_POSTFIX } from './common'
 
 export interface IEthersCfg {
   outDir?: string
@@ -32,11 +32,14 @@ export default class Ethers extends TypeChainTarget {
 
   private readonly allContracts: string[]
   private readonly outDirAbs: string
-  private readonly contractCache: Dictionary<{
-    abi: any
-    contract: Contract
-  }> = {}
-  private readonly bytecodeCache: Dictionary<BytecodeWithLinkReferences> = {}
+  private readonly contractCache: Dictionary<
+    | {
+        abi: any
+        contract: Contract
+      }
+    | undefined
+  > = {}
+  private readonly bytecodeCache: Dictionary<BytecodeWithLinkReferences | undefined> = {}
 
   constructor(config: Config) {
     super(config)
@@ -71,7 +74,7 @@ export default class Ethers extends TypeChainTarget {
     }
 
     if (this.contractCache[name]) {
-      const { contract, abi } = this.contractCache[name]
+      const { contract, abi } = this.contractCache[name]!
       delete this.contractCache[name]
       return [this.genContractFactoryFile(contract, abi, bytecode)]
     } else {
@@ -121,7 +124,7 @@ export default class Ethers extends TypeChainTarget {
     // For each contract that doesn't have bytecode (it's either abstract, or only ABI was provided)
     // generate a simplified factory, that allows to interact with deployed contract instances.
     const abstractFactoryFiles = Object.keys(this.contractCache).map((contractName) => {
-      const { contract, abi } = this.contractCache[contractName]
+      const { contract, abi } = this.contractCache[contractName]!
       return {
         path: join(this.outDirAbs, 'factories', `${contract.name}${FACTORY_POSTFIX}.ts`),
         contents: codegenAbstractContractFactory(contract, abi),
