@@ -1,2 +1,2 @@
-export { codegenArtifactHeaders } from './headers'
 export { codegenContract } from './contracts'
+export { codegenArtifactHeaders } from './headers'
