@@ -1,4 +1,4 @@
-import { AbiParameter, AbiOutputParameter, EvmType, EvmOutputType, TupleType } from 'typechain'
+import { AbiOutputParameter, AbiParameter, EvmOutputType, EvmType, TupleType } from 'typechain'
 
 export function codegenInputTypes(input: Array<AbiParameter>): string {
   if (input.length === 0) {
