@@ -1,7 +1,7 @@
 import BigNumber from 'bn.js'
-import { typedAssert, asyncWithDoneCase } from 'test-utils'
+import { asyncWithDoneCase, typedAssert } from 'test-utils'
 
-import { EventsInstance, Event3_bool_uint256 } from '../types/truffle-contracts/Events'
+import { Event3_bool_uint256, EventsInstance } from '../types/truffle-contracts/Events'
 
 const Events = artifacts.require('Events')
 
