@@ -1,3 +1,3 @@
-export * from './test'
 export * from './contract'
+export * from './test'
 export * from 'conditional-type-checks'
