@@ -1,9 +1,9 @@
 export * from './parser/abiParser'
-export * from './parser/parseEvmType'
-export * from './utils/files'
-export * from './utils/signatures'
-export * from './typechain/types'
-export * from './typechain/runTypeChain'
-export * from './utils/glob'
-export * from './typechain/types'
 export { normalizeName } from './parser/normalizeName'
+export * from './parser/parseEvmType'
+export * from './typechain/runTypeChain'
+export * from './typechain/types'
+export * from './typechain/types'
+export * from './utils/files'
+export * from './utils/glob'
+export * from './utils/signatures'
